Warning for ignored bf_categorize_counters()
bf_categorize_counters() does nothing in the absence of the
-bf-every-bb command-line option.  Hence, we now issue a run-time
warning message if this situation is detected and recommend that the
user specify -bf-every-bb -bf-merge-bb=18446744073709551615.

+++ lib/bytesflops/bytesflops.cpp
@@ -45,7 +45,7 @@ namespace bytesflops_pass {
 
   // Define a command-line option for merging basic-block measurements
   // to reduce the output volume.
-  cl::opt<int>
+  cl::opt<unsigned long long>
   BBMergeCount("bf-merge-bb", cl::init(1), cl::NotHidden,
                cl::desc("Merge this many basic blocks into a single line of output"),
                cl::value_desc("count"));
